requirements.txt
anthropic 
python-dotenv
json_repair